Refresh delegations when adding new delegation
