+++ web/src/components/Delegations/DelegationContext.tsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+export const DelegationContext = createContext({
+  refresh: () => {},
+});
