+++ web/src/components/Delegations/index.tsx
@@ -1,0 +1,3 @@
+export * from './RegisterDelegationKey';
+export * from './DelegationContext';
+export * from './Delegations';
