+++ web/src/components/index.tsx
@@ -1,5 +1,4 @@
 export * from './Form';
 export * from './NavBar';
-export * from './DelegationList';
-export * from './RegisterDelegateKey';
+export * from './Delegations';
 export * from './CreateTarget';
